+++ src/main/java/kr/hhplus/be/server/reservation/application/dto/ReservationResult.java
@@ -1,10 +1,8 @@
 package kr.hhplus.be.server.reservation.application.dto;
 
-import kr.hhplus.be.server.concert.domain.model.Seat;
 import kr.hhplus.be.server.reservation.domain.model.ReservationStatus;
 import lombok.Builder;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -13,7 +11,7 @@ public record ReservationResult(
         Long reservationId,
         LocalDateTime concertAt,
         Long seatId,
-        BigDecimal price,
+        Long price,
         ReservationStatus status,
         LocalDateTime expiredAt
